Fixed a bug in the autogen DMC for the digital inputs.
If 6 or more motors were configured to stop on the same IN, this will result
in a long program line in the generated DMC, the DMC will generate an error
when programmed with this generated file. This was checked with GalilTools
which showed much clear error message. The fix is simply a new line after
each axis.